requirements.txt
websockets>=12.0
python-dotenv>=1.0
folium>=0.15
pandas>=2.0
requests>=2.31
